refactor: Move test_kit from src/ to tests/ dir
test_kit is test-only tooling (imports vitest), not production library
code — doesn't belong under src/. Relocates to tests/test_kit/, wired via
TS project references (src/tsconfig.json composite: true, tests/test_kit's
own tsconfig referencing it) so dist/test_kit/index.js — and the package's
./test-kit export — stay unchanged. Zero impact on apify-mcp-server-internal,
which only ever imports the @apify/actors-mcp-server/test-kit subpath.
