+++ oxlint.config.ts
@@ -19,12 +19,9 @@ export default defineConfig({
             },
         },
         {
-            // src/test_kit/cases/*.cases.ts define `Case` objects whose `run(ctx)` is built via
-            // the `withClient(options, testFn)` helper — `expect()` calls inside `testFn` really
-            // do execute inside a real `it()` block (wired up by `registerCases` in register.ts),
-            // but oxlint's static vitest rules can't trace that indirection and flag them as
-            // standalone. False positive, not a real issue — see src/test_kit/register.ts.
-            files: ['src/test_kit/**'],
+            // Case.run() executes inside a real it() via registerCases (register.ts) — oxlint
+            // can't trace that indirection, flags below are false positives.
+            files: ['tests/test_kit/**'],
             rules: {
                 'vitest/no-standalone-expect': 'off',
                 'vitest/expect-expect': 'off',
