+++ package.json
@@ -103,11 +103,11 @@
     "start": "tsx src/dev_server.ts",
     "dev": "node scripts/dev_standby.js",
     "build": "pnpm run build:core && pnpm run build:web",
-    "build:core": "tsc -b src",
-    "build:watch": "tsc -b src -w",
+    "build:core": "tsc -b src && tsc -b tests/test_kit",
+    "build:watch": "tsc -b src -w & tsc -b tests/test_kit -w & wait",
     "build:web": "pnpm --filter @apify/mcp-web-widget run build",
     "postbuild": "mkdir -p dist/web && (cp -r src/web/dist dist/web/ 2>/dev/null || :)",
-    "clean": "tsc -b src --clean && rm -rf dist/web",
+    "clean": "tsc -b src --clean && tsc -b tests/test_kit --clean && rm -rf dist/web",
     "lint": "oxlint .",
     "lint:fix": "oxlint . --fix",
     "format": "oxfmt",
