+++ tests/test_kit/cases/actors.cases.ts
@@ -1,10 +1,10 @@
 import { expect } from 'vitest';
 
-import { ApifyClient } from '../../apify_client.js';
-import { HELPER_TOOLS, MAX_LIMIT_WITH_INPUT_SCHEMA } from '../../const.js';
-import { actorNameToToolName } from '../../tools/actor_tool_naming.js';
+import { ApifyClient } from '../../../src/apify_client.js';
+import { HELPER_TOOLS, MAX_LIMIT_WITH_INPUT_SCHEMA } from '../../../src/const.js';
+import { actorNameToToolName } from '../../../src/tools/actor_tool_naming.js';
 // Import tools from getCategoryTools instead of directly to avoid circular dependency during module initialization
-import { getCategoryTools } from '../../tools/index.js';
+import { getCategoryTools } from '../../../src/tools/index.js';
 import {
     ACTOR_EXAMPLE_MCP_SERVER,
     ACTOR_NORMAL_MODE,
