+++ tests/test_kit/cases/tools.cases.ts
@@ -1,8 +1,8 @@
 import type { StreamableHTTPClientTransport } from '@modelcontextprotocol/sdk/client/streamableHttp.js';
 import { expect } from 'vitest';
 
-import { HELPER_TOOLS } from '../../const.js';
-import { CALL_ACTOR_MCP_MISSING_TOOL_NAME_MSG } from '../../tools/actors/call_actor.js';
+import { HELPER_TOOLS } from '../../../src/const.js';
+import { CALL_ACTOR_MCP_MISSING_TOOL_NAME_MSG } from '../../../src/tools/actors/call_actor.js';
 import {
     ACTOR_EXAMPLE_MCP_SERVER,
     asLegacyClient,
