+++ tests/test_kit/register.ts
